+++ packages/pipeline/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.6.32 (2026-02-16)
+
+### 🧱 Updated Dependencies
+
+- Updated @lde/dataset-registry-client to 0.6.17
+- Updated @lde/local-sparql-endpoint to 0.2.12
+- Updated @lde/sparql-importer to 0.2.10
+- Updated @lde/sparql-server to 0.4.10
+- Updated @lde/dataset to 0.6.10
+
+## 0.6.31 (2026-02-16)
+
+### 🚀 Features
+
+- **pipeline:** add SparqlServer support to distribution resolver ([#118](https://github.com/ldengine/lde/pull/118))
+
+### ❤️ Thank You
+
+- David de Boer
+
 ## 0.6.30 (2026-02-16)
 
 ### 🩹 Fixes