+++ packages/wait-for-sparql/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.11 (2026-02-16)
+
+This was a version bump only for @lde/wait-for-sparql to align it with other projects, there were no code changes.
+
 ## 0.2.10 (2026-02-13)
 
 This was a version bump only for @lde/wait-for-sparql to align it with other projects, there were no code changes.